pyproject.toml
[project]
name = "audiblecontrol"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "kivy>=2.3.1",
    "openai>=1.92.2",
    "pyaudio>=0.2.14",
    "pyserial>=3.5",
    "python-dotenv>=1.1.1",
    "setuptools-rust>=1.11.1",
    "sounddevice>=0.5.2",
    "speechrecognition>=3.14.3",
]
